feat: Add option to render raw HTML in column values
This change introduces a new `render_raw` option in the column settings, which determines whether the column value should be rendered as raw HTML. This allows for more flexibility in displaying data in our tables when needed, particularly in cases where HTML tags within the data should be displayed as is.

The diff reflects changes to base.html.twig to add conditional rendering of column values based on the `render_raw` value. The option was also added as a boolean in ColumnType.php, with its default set to false. Documentation was updated accordingly.